MDEV-29855 Crash with SPIDER_DIRECT_SQL and spider_udf_ds_use_real_table=1
The crash occurs because of the following call of TABLE_LIST::init_one_table():

  table_list.init_one_table(
    &table_list.db, &table_list.table_name, 0, TL_WRITE);

One should not pass table_list.db and table_list.table_name to the function
because it update the very members internally.

The function is called previously, and there is no need to call it again.
So, simply removing the call will resolve the problem.